fix: tiny fix cli
