+++ python/sglang/cli/serve.py
@@ -4,7 +4,7 @@ import argparse
 import logging
 import os
 
-from sglang.cli.main import get_is_diffusion_model, get_model_path
+from sglang.cli.utils import get_is_diffusion_model, get_model_path
 from sglang.srt.utils import kill_process_tree
 
 logger = logging.getLogger(__name__)
